Remove useless open mode
Open modes are ignored if the user doesn't want the file to be created:
http://man7.org/linux/man-pages/man2/open.2.html

Since it's somewhat misleading/potentially indicative of a missing
O_CREAT/O_TMPFILE, we plan to issue a warning on open calls with useless
mode bits in the near future. Because -Werror is used in this project
(thank you!), we need to either make the mode useful, or remove it.

Bug: 64132680
Test: mma. New warning no longer appears.
Change-Id: I6d9e59355a0071bc2989f54a3b921a36b90d1857