kselftest_x86_tests: require libdl.a explicitly
libdl.a provides dummy no-op implementations of dynamic loader functions
(e.g. dlopen, dladdr) for static executables. Previously, libdl.a was
linked automatically into static executables, but now the library is
opt-in.

The unwind_vdso.c test needs the dummy library, because it calls dladdr to
print the filename and base address of the vDSO. Because the test is a
static executable, dladdr will return 0, and the test will print
"[WARN]\tdladdr failed on AT_SYSINFO\n".

Bug: b/141485154
Test: m device_kselftest_x86_tests_unwind_vdso_all_targets
Change-Id: I60878459f70a35aab8b7b02914fd84a4164f2d12