Disable native_coverage for kselftest_size_test
Bug: http://b/156301567

Coverage instrumentation is infeasible with nocrt - Clang's coverage
runtime calls atexit, which is provided by crtbegin_dynamic.o.  Disable
coverage for just this test.  Doing this at the module level so we know
about this consequence of 'nocrt' instead of silently doing so in the
build system.  Also, 'nocrt' is an uncommon setting so this won't be too
noisy.

Test: m CLANG_COVERAGE=true \
      COVERAGE_PATHS="external/linux-kselftest" \
      external/linux-kselftest
Change-Id: Ia964321d031d48bba6f96ef2ed961ca979aae90b