FROMLIST: selftests: futex: Use a 64-bit time_t
Convert the futex selftests to only use a 64-bit time_t. On 64-bit
architectures this isn't a functional change. On 32-bit architectures
we now only perform 64-bit time_t syscalls (__NR_futex_time64) and
use a struct timespec64.

This won't work on kernels before 5.1, but as perf is tied to the kernel
that's ok.

This allows the tests to run and pass on RISC-V 32-bit.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>
Reviewed-by: Arnd Bergmann <arnd@arndb.de>
Change-Id: Ic61bbedf667ea8d93928916a39c924333cb10c28
Link: https://lore.kernel.org/all/20211014055527.1238645-2-alistair.francis@opensource.wdc.com/
Bug: 305651744
Test: atest -a vts_linux_kselftest_x86_32:futex_functional_run.sh_x86_32
Test: atest -a vts_linux_kselftest_x86_64:futex_functional_run.sh_x86_64
Signed-off-by: Edward Liaw <edliaw@google.com>