ANDROID: kvm: Test that pVM memory is wiped during teardown
In protected KVM mode, we expect the hypervisor to protect guest secrets
when they are torn down. Add a test checking this property by running a
minimal guest, and checking that the content of memory has been wiped by
the hypervisor after teardown.

Note: although some of the pKVM code has already landed upstream, the
functionality tested here hasn't at the time of writing. Once it does,
this test should be sent upstream for review to replace this ANDROID
patch.

Bug: 218934075
Signed-off-by: Quentin Perret <qperret@google.com>
Change-Id: I4a347908d189f2c4835fd576b26fae7c10ec9b22