userfaultfd: infinite loop in faulting_process
The local variables steps and signalled were being reset to 1 and 0
respectively after every jump back to sigsetjmp by siglongjmp in the
signal handler. The test was incrementing them and expecting them to
retain their incremented values. The documentation for siglongjmp says:

All accessible objects have values as of the time sigsetjmp() was
called, except that the values of objects of automatic storage duration
which are local to the function containing the invocation of the
corresponding sigsetjmp() which do not have volatile-qualified type and
which are changed between the sigsetjmp() invocation and siglongjmp()
call are indeterminate.

Tagging steps and signalled with volatile enabled the test to pass.

Bug: 232026677
Signed-off-by: Edward Liaw <edliaw@google.com>
Change-Id: I03242a008bb2672431a441e9fa1e7564779382ce